requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.0.200
pytest>=7.4.0
pytest-django>=4.5.2
psycopg2-binary>=2.9.6
tzdata>=2024.1
langchain_openai>=0.3.16
google-cloud-vision>=3.0.2
boto3==1.38.22